backend/requirements.txt
openai==1.12.0
opencv-python==4.9.0.80
requests==2.31.0
python-dotenv==1.0.1
flask==3.0.2